* fix_fdat_size:
  update ChangeLog
  fix, Fx.dat file might be larger then needed
  add error handling to FDat::countRecords()
  FDat::print(), use the size record instead of file size
  FDat, use singed types for countRecords() and record_length
  style, cleanup master's countRecords() functions
  allow not aligned  master file sizes
  style, remove usless outcommented assert

+++ src/ms_file.h
@@ -239,7 +239,7 @@ class FDat
 		
 		bool checkHeader() const;
 		int print( const char* header ) const;
-		unsigned short countRecords() const;
+		int countRecords() const;
 		
 	private:
 		static int header_to_string( char *s );
@@ -254,7 +254,7 @@ class FDat
 		static ftoa_func opi_ftoa;
 		
 		const unsigned char field_bitset;
-		const unsigned int record_length;
+		const int record_length;
 		
 		const char * const buf;
 		const int size;
